requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
opencv-python-headless
Pillow
streamlit
requests
numpy
scikit-learn